+++ packages/core-js-compat/src/data.mjs
@@ -1973,6 +1973,10 @@ export const data = {
   },
   'esnext.composite-symbol': {
   },
+  'esnext.data-view.get-uint8-clamped': {
+  },
+  'esnext.data-view.set-uint8-clamped': {
+  },
   'esnext.disposable-stack.constructor': {
   },
   'esnext.function.demethodize': {