+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -213,4 +213,8 @@ export default {
     'web.url-search-params.delete',
     'web.url-search-params.has',
   ],
+  3.32: [
+    'esnext.data-view.get-uint8-clamped',
+    'esnext.data-view.set-uint8-clamped',
+  ],
 };